last upload code

+++ app/src/main/java/com/infinity/mental/data/model/result/CfKombinasi.kt
@@ -1,6 +1,6 @@
 package com.infinity.mental.data.model.result
 
 data class CfKombinasi(
-    val cf: List<Int>,
+    val cf: List<Double>,
     val kode_depresi: List<String>
 )
